ARGO-3445 Refactor packages of ArgoStatusBatch job
