+++ flink_jobs_v2/batch_status/src/main/java/trends/calculations/CalcMetricFlipFlopTrends.java
@@ -1,4 +1,4 @@
-package flipflops;
+package trends.calculations;
 
 /*
  * To change this license header, choose License Headers in Project Properties.
