+++ flink_jobs_v2/batch_status/src/main/java/trends/calculations/CalcServiceFlipFlopTrends.java
@@ -1,14 +1,14 @@
-package flipflops;
+package trends.calculations;
 
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-//import argo.functions.calctimelines.TimelineMerger;
 import argo.batch.StatusTimeline;
 import argo.batch.TimeStatus;
 import java.util.ArrayList;
+
 import java.util.HashMap;
 import java.util.TreeMap;
 import org.apache.flink.api.common.functions.FlatMapFunction;
@@ -22,18 +22,20 @@ import timelines.Timeline;
  * CalcServiceEndpointFlipFlop, count status changes for each service endpoint
  * group
  */
-public class CalcGroupFlipFlopTrends implements FlatMapFunction<StatusTimeline, GroupTrends> {
+public class CalcServiceFlipFlopTrends implements FlatMapFunction<StatusTimeline, ServiceTrends> {
 
-    public CalcGroupFlipFlopTrends() {
+    public CalcServiceFlipFlopTrends() {
+
     }
+    static Logger LOG = LoggerFactory.getLogger(CalcServiceFlipFlopTrends.class);
 
-    static Logger LOG = LoggerFactory.getLogger(CalcGroupFlipFlopTrends.class);
-   
     @Override
-    public void flatMap(StatusTimeline in, Collector<GroupTrends> out) throws Exception {
-     String group = in.getGroup();  
-     
-   ArrayList<TimeStatus> timestatusList = in.getTimestamps();
+    public void flatMap(StatusTimeline in, Collector<ServiceTrends> out) throws Exception {
+
+        String group = in.getGroup();
+        String service = in.getService();
+        //construct a timeline containing all the timestamps of each metric timeline
+        ArrayList<TimeStatus> timestatusList = in.getTimestamps();
 
         TreeMap<DateTime, Integer> timestampMap = new TreeMap();
         for (TimeStatus ts : timestatusList) {
@@ -46,12 +48,10 @@ public class CalcGroupFlipFlopTrends implements FlatMapFunction<StatusTimeline, 
         timelineMap.put("timeline", timeline);
         Integer flipflop = timeline.calcStatusChanges();
 
-        if (group != null ) {
-
-            GroupTrends groupTrends = new GroupTrends(group, timeline, flipflop);
-            out.collect(groupTrends);
+        if (group != null && service != null) {
+            ServiceTrends serviceTrends = new ServiceTrends(group, service, timeline, flipflop);
+            out.collect(serviceTrends);
         }
-
     }
 
 }
