+++ flink_jobs_v2/batch_status/src/main/java/trends/calculations/EndpointTrends.java
@@ -1,39 +1,56 @@
-package flipflops;
+package trends.calculations;
 
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-
 import timelines.Timeline;
 
 /**
- * MetricTrends, describes the computed trend information extracted from the set
- * of the timelines at the level of group service endpoints metrics groups *
+ * EndpointTrends, describes the computed trend information extracted from the
+ * timelines at the level of group service endpoints groups
  */
-public class GroupTrends {
+public class EndpointTrends {
 
     String group;
-
+    String service;
+    String endpoint;
     Timeline timeline;
     Integer flipflops;
 
-    public GroupTrends() {
+    public EndpointTrends() {
     }
 
-    public GroupTrends(String group, Timeline timeline, Integer flipflops) {
+    public EndpointTrends(String group, String service, String endpoint, Timeline timeline, Integer flipflops) {
         this.group = group;
+        this.service = service;
+        this.endpoint = endpoint;
         this.timeline = timeline;
         this.flipflops = flipflops;
     }
-
     public String getGroup() {
         return group;
     }
 
     public void setGroup(String group) {
         this.group = group;
+    }
+
+    public String getService() {
+        return service;
+    }
+
+    public void setService(String service) {
+        this.service = service;
+    }
+
+    public String getEndpoint() {
+        return endpoint;
+    }
+
+    public void setEndpoint(String endpoint) {
+        this.endpoint = endpoint;
     }
 
     public Timeline getTimeline() {
