+++ flink_jobs_v2/batch_status/src/main/java/trends/calculations/GroupTrends.java
@@ -1,33 +1,29 @@
-package flipflops;
+package trends.calculations;
 
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+
 import timelines.Timeline;
 
 /**
  * MetricTrends, describes the computed trend information extracted from the set
- * of the timelines at the level of group service endpoints metrics groups
+ * of the timelines at the level of group service endpoints metrics groups *
  */
-public class MetricTrends {
+public class GroupTrends {
 
     String group;
-    String service;
-    String endpoint;
-    String metric;
+
     Timeline timeline;
     Integer flipflops;
 
-    public MetricTrends() {
+    public GroupTrends() {
     }
 
-    public MetricTrends(String group, String service, String endpoint, String metric, Timeline timeline, Integer flipflops) {
+    public GroupTrends(String group, Timeline timeline, Integer flipflops) {
         this.group = group;
-        this.service = service;
-        this.endpoint = endpoint;
-        this.metric = metric;
         this.timeline = timeline;
         this.flipflops = flipflops;
     }
@@ -38,30 +34,6 @@ public class MetricTrends {
 
     public void setGroup(String group) {
         this.group = group;
-    }
-
-    public String getService() {
-        return service;
-    }
-
-    public void setService(String service) {
-        this.service = service;
-    }
-
-    public String getEndpoint() {
-        return endpoint;
-    }
-
-    public void setEndpoint(String endpoint) {
-        this.endpoint = endpoint;
-    }
-
-    public String getMetric() {
-        return metric;
-    }
-
-    public void setMetric(String metric) {
-        this.metric = metric;
     }
 
     public Timeline getTimeline() {
@@ -79,4 +51,5 @@ public class MetricTrends {
     public void setFlipflops(Integer flipflops) {
         this.flipflops = flipflops;
     }
+
 }
