+++ flink_jobs_v2/batch_status/src/main/java/trends/calculations/MongoTrendsOutput.java
@@ -1,4 +1,4 @@
-package flipflops;
+package trends.calculations;
 
 import java.io.IOException;
