+++ flink_jobs_v2/batch_status/src/main/java/trends/calculations/ServiceTrends.java
@@ -1,34 +1,34 @@
-package flipflops;
-
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+package trends.calculations;
+
 import timelines.Timeline;
 
 /**
- * EndpointTrends, describes the computed trend information extracted from the
+ * ServiceTrends, describes the computed trend information extracted from the
  * timelines at the level of group service endpoints groups
  */
-public class EndpointTrends {
+public class ServiceTrends {
 
     String group;
     String service;
-    String endpoint;
     Timeline timeline;
     Integer flipflops;
+    String function;
 
-    public EndpointTrends() {
+    public ServiceTrends() {
     }
 
-    public EndpointTrends(String group, String service, String endpoint, Timeline timeline, Integer flipflops) {
+    public ServiceTrends(String group, String service, Timeline timeline, Integer flipflops) {
         this.group = group;
         this.service = service;
-        this.endpoint = endpoint;
         this.timeline = timeline;
         this.flipflops = flipflops;
     }
+
     public String getGroup() {
         return group;
     }
@@ -45,14 +45,6 @@ public class EndpointTrends {
         this.service = service;
     }
 
-    public String getEndpoint() {
-        return endpoint;
-    }
-
-    public void setEndpoint(String endpoint) {
-        this.endpoint = endpoint;
-    }
-
     public Timeline getTimeline() {
         return timeline;
     }
@@ -67,6 +59,14 @@ public class EndpointTrends {
 
     public void setFlipflops(Integer flipflops) {
         this.flipflops = flipflops;
+    }
+
+    public String getFunction() {
+        return function;
+    }
+
+    public void setFunction(String function) {
+        this.function = function;
     }
 
 }
