+++ flink_jobs_v2/batch_status/src/main/java/trends/flipflops/MapGroupTrends.java
@@ -1,5 +1,7 @@
-package flipflops;
+package trends.flipflops;
 
+import trends.calculations.GroupTrends;
+import trends.calculations.Trends;
 import org.apache.flink.api.common.functions.MapFunction;
 
 public class MapGroupTrends implements MapFunction<GroupTrends, Trends> {
