+++ flink_jobs_v2/batch_status/src/main/java/trends/flipflops/MapServiceTrends.java
@@ -1,4 +1,6 @@
-package flipflops;
+package trends.flipflops;
+import trends.calculations.ServiceTrends;
+import trends.calculations.Trends;
 import org.apache.flink.api.common.functions.MapFunction;
 
 public class MapServiceTrends implements MapFunction<ServiceTrends, Trends> {
