+++ flink_jobs_v2/batch_status/src/main/java/trends/flipflops/ZeroMetricFlipFlopFilter.java
@@ -1,9 +1,10 @@
-package flipflops;
+package trends.flipflops;
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
+import trends.calculations.MetricTrends;
 import org.apache.flink.api.common.functions.FilterFunction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -11,9 +12,9 @@ import org.slf4j.LoggerFactory;
 /**
 * ZeroMetricTrendsFilter, filters MetricTrends data and rejects the ones with flipflop=0
 */
-public class ZeroMetricTrendsFilter implements FilterFunction<MetricTrends> {
+public class ZeroMetricFlipFlopFilter implements FilterFunction<MetricTrends> {
 
-    static Logger LOG = LoggerFactory.getLogger(ZeroMetricTrendsFilter.class);
+    static Logger LOG = LoggerFactory.getLogger(ZeroMetricFlipFlopFilter.class);
 
    
     //if the status field value in Tuple equals the given status returns true, else returns false
