+++ flink_jobs_v2/batch_status/src/main/java/trends/flipflops/ZeroServiceFlipFlopFilter.java
@@ -1,32 +1,29 @@
-package flipflops;
-
-
-
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-
+package trends.flipflops;
+import trends.calculations.ServiceTrends;
 import org.apache.flink.api.common.functions.FilterFunction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /**
- * ZeroEndpointTrendsFilter, filters EndpointTrends data and rejects the ones with flipflop=0
+ * ZeroServiceTrendsFilter, filters ServiceTrends data and rejects the ones with flipflop=0
  */
-public class ZeroEndpointTrendsFilter implements FilterFunction<EndpointTrends> {
-
-    static Logger LOG = LoggerFactory.getLogger(ZeroEndpointTrendsFilter.class);
+    public class ZeroServiceFlipFlopFilter implements FilterFunction<ServiceTrends> {
+    
+    static Logger LOG = LoggerFactory.getLogger(ZeroServiceFlipFlopFilter.class);
 
    
     //if the status field value in Tuple equals the given status returns true, else returns false
     @Override
-    public boolean filter(EndpointTrends t) throws Exception {
+    public boolean filter(ServiceTrends t) throws Exception {
         if (t.getFlipflops()>0) {
             return true;
         }
         return false;
     }
-
+    
 }
