+++ flink_jobs_v2/batch_status/src/main/java/trends/status/MetricTrendsCounter.java
@@ -1,11 +1,11 @@
-package trends;
+package trends.status;
 
 /*
  * To change this license header, choose License Headers in Project Properties.
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-import flipflops.MetricTrends;
+import trends.calculations.MetricTrends;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
